Introduce custom ViewDesc implementations.
The ViewDesc classes are where prosemirror-view actually does all of
its rendering work. We are introducing stubs of these classes that
reproduce all of the functionality of the originals, but do not provide
update or destroy methods, which means that they do not actually do any
DOM manipulation.

The goal is to later introduce React components that replicate the
rendering functionality of the original ViewDesc classes.