Add some hooks to replace EditorView functionality
The useBeforeInput hook replaces the DOMObserver functionality that
we plan to remove from EditorView.

usePluginViews manages plugin view updates, to ensure that they occur
after the DOM has been updated to match the state.

useSyncSelection similarly ensures that the ProseMirror selection is
synced to the DOM after the DOM has been updated to match the state.