Add a selection-only DOMObserver.
EditorView's DOMObserver uses a Mutation Observer to detect changes to
the contenteditable, and then attempts to deduce the intended change to
state from that. This introduces challenges when applied to React's
unidirection data flow model.

To work around this, we use a beforeinput handler to detect changes
before they're committed to the DOM. This means our DOMObserver only
needs to worry about syncing the selection, so we added a custom one
that only has that concern, but matches the API of the original
implementation.