Map positions forward to maintain key stability
Rather than attempting to map the new node positions backward to
determine what the key used to be for that node, start with the previous
positions and map them forward, taking care to skip any positions that
are deleted by the transaction, and adding any new nodes that didn't
exist in the old doc. This improves stability and correctness in more
complex cases, such as splitting or wrapping nodes, and nested
deletions.