Use consistent module layout and naming
Move every module in the contexts directory to the components directory.
Name each one with a consistent suffix "Context" suffix. From each one,
export the context and the context value interface.

For every context module, ensure there is a corresponding module with a
component that wraps the provider. These modules do not have any suffix.

Move the useLayoutGroupEffect hook into the hooks directory. Fix all of
the remaining references to the old "deferred layout effects" naming.

Export only a single EditorProps type that is the type of the props for
the ProseMirror and Editor components. Hooks do not export their types.

Make all of the exports in the barrel module consistent.

Rename PortalRegistry to NodeViews.