Use consistent module layout and naming (#97)
For every context module, ensure there is a corresponding module with a
component that wraps the provider.

Move the useLayoutGroupEffect hook into the hooks directory. Fix all of
the remaining references to the old "deferred layout effects" naming.

Export only a single ProseMirrorProps type that is the type of the props for
the ProseMirror component. Hooks do not export their types.

Make all of the exports in the barrel module consistent.

Rename PortalRegistry to NodeViews.

Stop exporting the low-level building blocks of the main API