Make proper use of view.someProp
There were a few places where we incorrectly used view.someProp to
retrieve a prop value, and then did something with the value. This will
_always_ result in the first value being returned, even in cases where
multiple plugins provide a value for the prop, and only some of them
are relevant at the callsite.

Instead, we now pass a function to view.someProp that only returns a
truthy value when it has found a prop provider that provides the
relevant value, so that we can fallback through the entire list of
plugins.